+++ backend/src/main/java/cz/itnetwork/dto/PersonRevenueDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class PersonRevenueDTO {
     private Long personId;
     private String name;
     private String identificationNumber;
-    private long revenue;
+    private BigDecimal revenue;
 }
